+++ src/main/java/org/kumoricon/model/order/OrderRepository.java
@@ -14,4 +14,7 @@ public interface OrderRepository extends JpaRepository<Order, Integer> {
     @Query(value = "SELECT paymentType, COUNT(id) as cnt, SUM(totalAmount) as total FROM orders WHERE paymentTakenByUser_id = ?1 AND paidSession = ?2 GROUP BY paymentType;", nativeQuery = true)
     List<Object[]> getSessionOrderCountsAndTotals(Integer userId, Integer sessionNumber);
 
+    @Query(value = "SELECT SUM(totalAmount) as total FROM orders WHERE paymentTakenByUser_id = ?1 AND paidSession = ?2 ;", nativeQuery = true)
+    Float getSessionOrderTotal(Integer userId, Integer sessionNumber);
+
 }