+++ CHANGES.md
@@ -9,6 +9,8 @@ Change Log
     - Rearranged fields on User edit window to be more convenient; sername and badge prefix will be auto generated
     - Added configuration option to display "training mode" banner behind left menu
     - Log IP address in place of username for users that aren't logged in when they view a page
+    - Added a prompt when logging out if user needs to close their till (has orders worth more than $0 
+      in current session)
 
 - Version 0.9.5 (08/20/2016)
     - Attendee importer will work when there are only 18 fields instead of 19 (notes field is missing/blank)